added extension functionality
