+++ index.json
@@ -1,0 +1,22 @@
+{
+  "version": "v1",
+  "blocklist": [],
+  "data": [
+    {
+      "schema_version": "1.0.0",
+      "id": "woody",
+      "name": "Woody",
+      "tagline": "A piepline tool",
+      "version": "0.2.1",
+      "type": "add-on",
+      "maintainer": "ITB",
+      "license": [
+        "SPDX:GPL-3.0-or-later"
+      ],
+      "blender_version_min": "4.2.0",
+      "archive_url": "./woody.zip",
+      "archive_size": 22223,
+      "archive_hash": "sha256:d76bc491bb1605b7e677597c440027656d526a8a8bb78bde8b5883696b38d9e2"
+    }
+  ]
+}
