+++ woody/blender_manifest.toml
@@ -1,0 +1,18 @@
+schema_version = "1.0.0"
+
+id = "woody"
+version = "0.2.1"
+name = "Woody"
+tagline = "A pipeline tool"
+maintainer = "ITB"
+
+type = "add-on"
+
+blender_version_min = "4.2.0"
+
+license = [
+  "SPDX:GPL-3.0-or-later",
+]
+
+files = "Import/export"
+clipboard = "Copy and paste"
